cleans up routes
