+++ app/routes/routes.py
@@ -80,30 +80,6 @@ def users():
     trips = db.session.execute(db.select(Trip).join_from(User, Trip).where(User.username == username)).scalars()
     return render_template('trips.html', trips=trips)
 
-@app.route('/experiences', methods=['GET'])
-def experiences():
-    search = request.args.get('q')
-    sort = request.args.get('sort')
-    
-    if search:
-        experiences = Experience.query.filter(or_(Experience.title.contains(search), Experience.description.contains(search)))
-    else:
-        experiences = Experience.query
-
-    if sort == 'highest_rating':
-        experiences = experiences.order_by(Experience.rating.desc())
-    elif sort == 'lowest_rating':
-        experiences = experiences.order_by(Experience.rating.asc())
-
-    experiences = experiences.all()
-
-    return render_template('experiences.html', experiences=experiences)
-
-@app.route('/experience/<int:experience_id>', methods=['GET'])
-def experience_detail(experience_id):
-    experience = Experience.query.get_or_404(experience_id)
-    return render_template('experience_detail.html', experience=experience)
-
 @app.route('/health')
 def health():
     return jsonify({'status': 'ok'}), 200
@@ -112,7 +88,3 @@ def health():
 @app.route('/index')
 def index():
     return render_template('index.html')
-
-
-
-
